backend/requirements.txt
amqp==5.3.1
asgiref==3.11.0
async-timeout==5.0.1
billiard==4.2.4
cachetools==6.2.2
celery==5.6.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==2.0.6
django==5.2.8
django-autoslug==1.9.9
django-celery-beat==2.8.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-debug-toolbar==6.1.0
django-environ==0.12.0
django-phonenumber-field==8.4.0
django-redis==6.0.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-timezone-field==7.2.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
dnspython==2.8.0
drf-yasg==1.21.11
email-validator==2.3.0
exceptiongroup==1.3.1
google-auth==2.43.0
idna==3.11
inflection==0.5.1
itypes==1.2.0
jinja2==3.1.6
kombu==5.6.1
markupsafe==3.0.3
openapi-codec==1.3.2
packaging==25.0
phonenumbers==9.0.19
pillow==12.0.0
prompt-toolkit==3.0.52
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1-modules==0.4.2
pyjwt==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.3
redis==7.1.0
requests==2.32.5
rsa==4.9.1
simplejson==3.20.2
six==1.17.0
sqlparse==0.5.4
typing-extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
